pyproject.toml
[tool.poetry]
name = "script-news-for-discord-sraper"
version = "0.1.0"
description = ""
authors = ["tyronejose <alt.tyronejose@gmail.com>"]
license = "MIT Licence"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
colorlog = "^6.9.0"
beautifulsoup4 = "^4.12.3"
supabase = "^2.10.0"
pytrends = "^4.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
